step-2: разделил один большой редьюсер на пачку мелких
