+++ src/components/BurgerIngredientsItem/BurgerIngredientsItem.js
@@ -10,8 +10,8 @@ import { useSelector } from "react-redux";
 import { useDrag } from "react-dnd";
 
 const BurgerIngredientsItems = ({ data, onClick }) => {
-    const ingredients = useSelector(
-        (store) => store.burger.constructorIngredients,
+    const ingredients = useSelector(//x
+        (store) => store.burger.constructorReducer.ingredients,
     ).filter((item) => {
         return item && item._id === data._id;
     });
